MessagesFile: Remove uncolored strings
This methods were making the code way bigger than it should be, let's be honest, it is pretty unlikely for a plugin to actually want to send a configurable message without allowing administrators to add color patterns to it, this methods made almost to no sense on a class designed to send custom messages with support for fancy features. This also makes the "colored" part on the name unnecessary, so we have shorter names!