Replacer: Optimizations, num support always on
With this optimizations Replacers no longer need regular expressions for numeric support to work, meaning that performance is greatly improved and we no longer need to compile and store the Pattern, also saving a tiny bit of memory there. This commit also includes optimizations not related to numeric support, for example, replaceAt no longer uses Strings but a StringBuilder, avoiding the creation of multiple String instances.