Added FileHolder and FileUpdater (See commit desc)
This change has been made in order to allow MCPlugins to register files that aren't .yml files, this makes the code more scalable for the future if we add support for .json files or something like that. This is also useful for plugins right now that may use any file other type not supported by MCUtils.