Added BaseMCCommand#hasAccess (See desc)
This method has been added just in case any plugin wants to add its custom MCCommand or MCSubCommand class with any requirements for any command types that may be added, for example, a StaffCommand class that restricts access only to senders with a specific permission, removing code duplication and possible bugs like forgetting to do this permission check on one command. MCCommand and MCSubCommand do not override this method so it will be always true for them, meaning that this method is reserved for sub classes and that behavior must NOT be changed by MCUtils, never.