Added the Replacement interface (See desc)
I'm not really sure about the name of this interface as it's quite generic, however, Replacer is also generic so... I think it's ok for now to leave it like this. The Replacement interface will come handy for any object that may be added as a replacement on a replacer without having to call any method, something that was already done with players, that can be added to a Replacer and the Replacer will be "intelligent" enough to add their #getName method to the list instead of #toString.