MessagesFile: Minor improvements (See desc)
Default replacer usage has been optimized, previously the default replacer was being cloned twice, which obviously isn't ideal. Some readability improvements were also added in addition to the usage of the "this" keyword when calling methods from the MessagesFile. This is to avoid unexpected behavior on classes that extend MessagesFile and override any of those methods. This will also be added to PluginFile and YmlFile on a future commit.