MCCommand: Added tab complete filtering
This commit adds tab complete filtering to MCCommands. This is easier to understand with an example, like the /gamemode command, which on its first argument will return "adventure", "creative", "survival" and "spectator" as options. With tab complete filtering, if the user has already written "/gamemode C", the only suggestion that will be made to the user is "creative". This ignores caps.