Update package to net.codersky (See desc)
After quite some time thinking about this change I finally decided to do it now before the project develops any further. In reality, this could be considered an irrelevant change as I'm the owner of the Codersky Network and hence MCUtils still belongs to me. However, internally, this makes sense for me as the project is no longer linked to myself in case someone else contributes to it and the costs are lowered as I no longer have to maintain the xdec0de.me domain, only the codersky.net domain. This, however, means that all the package names have changed and the documentation needs to be changed, in my opinion, right now is the right moment to do this change before any plugin from any other author starts to use MCUtils and gets annoyed by this quite huge change.