MCStrings: Added asNumberFormat (See desc)
Alright so this one is a little bit tricky. I ended up doing the multiplications manually with a StringBuilder. This is because of rounding errors that occurred with the 'b' character or integers, so for example "2.1b". A lot of casting was involved just to support the generic Number class to avoid code duplication, so I ended up doing the operations with the string itself, as the multiplications only added zeros or removed the '.' character. There is currently a bug that will most likely not be fixed as it only happens with a pretty uncommon format "1.2345k", the result should be 1234.5, but the decimal part gets removed. I feel like the weird formatting itself defeats the purpose of this method which is to simplify let's say /pay operations such as "/pay user 2k", not to cover every weird user input.