Added the MessagesFileHolder interface (See desc)
This has been done because MessagesFile was the only class available to send messages directly from a file, and it is that, a class with an existing implementation that contains a default Replacer. While useful, not every plugin may want every feature from MessagesFile and may want to create its own custom MessagesFile type.

This makes it so you can now create your own MessagesFile class with your own implementation while also having basic yet useful methods with a default implementation similar to the behavior of MessagesFile but without a default replacer.