MCPlugin#getMessages: return MessagesFileHolder
As explained on the previous commit, this now allows plugins to return their own MessagesFileHolder types in a more convenient way than just extending MessagesFile (And being forced to have a default Replacer). Plugins are now also no longer limited to .yml files for messages because of this, which is always a good plus.

Additionally, this reduces confusion as of why the MCPlugin#getMessages method returns an object capable of returning things that aren't messages, like booleans, even though the MessagesFile class can still be used as a fully functional YmlFile, the MessagesFileHolder interface is limited to Strings, String lists and sending messages to CommandSenders, which makes more sense to the end user and makes clear what methods should be used and are actually implemented.