Replace ActionGUI with GUIActionMap (See desc)
This change was made mostly because using ActionGUI was a compromise as you were required to extend an abstract class, meaning that you wouldn't be able to extend other classes, this reduced flexibility. Now, GUIActionMap can be integrated on a regular GUI as a field, then used on GUI#onClick(Player, Inventory, InventoryClickEvent). ActionGUI can of course be recreated by other plugins by overriding this method, however, that is something I won't add to MCUtils as doing this is very basic (And will be present on the guide as a copy-paste template).