Added a way more flexible UpdateChecker (See desc)
MCPlugin#getLatestVersion(long) has been removed in favor of an UpdateChecker that can manage as many sources to get updates from as we want. Right now, we only will need to implement the new UpdaterSource interface and that will work as a new source to get the latest version of a plugin. Expect new sources such as github releases soon.