UpdateChecker: Added GitHubUpdaterSource
I don't really like the implementation but it should work. It currently gets the name of the release, I may end up changing that to the tag_name, as it seems like tags usually contain just the version, names may contain a title for the release, not the version. I'll do some research on a few repositories to decide. I could also use the jar itself, but that usually contains the name of the plugin too.