getLatestVersion returns VersionInfo (See desc)
This new change makes it so every UpdaterSource can return a VersionInfo object that contains more information than just the version name. HangarUpdaterSource has been updated in consequence, SpigotUpdaterSource will be upated soon.

For contributors:
You may have noticed that VersionInfo doesn't set the UpdaterSource field on its constructor, this is because of the way Gson instantiates classes, I didn't find a way to pass arguments, so #getLatestVersion() now sets the info.source field to this.