Warn checking for updates synchronously (See desc)
Both UpdateChecker and UpdaterSource now have an optional boolean parameter on their getLatestVersion method to ignore this warning. This warning exists just in case a developer forgets to get the latest version of a plugin asynchronously, as this would cause a VERY BIG performance issue outside of the JavaPlugin#onEnable method (Like a command to check for updates).