StorageHandler: Remove generic usage (See desc)
This commit adds quite a lot of methods, setters for each type and list of types that StorageHandler supports, as well as lists getters, hiding the generic method that was previously used. This change has been done because the types that StorageHandler supported wasn't clear, so you could for example try to store or get a Player instead of the UUID of the Player, leading to confusion. Now it is pretty clear what types are supported by StorageHandler as specific methods exist for each type.