FlatStorage: File breaking fixes
This fixes the usage of \n characters in strings and string lists, as well as trying to save empty lists, which happened when a list was obtained and then emptied, as StorageHandler doesn't return a copy of the stored list but the actual list, this is intentional but makes possible clearing the list. StorageHandler does indeed remove empty lists from the cache, but this is only done if a setXs method is called, not if the list is emptied externally, meaning that some lists may be empty when save is called, and we don't want to store empty lists.