MCCommand#asEnum: Always use upper case args
As the method documentation now mentions, enum constants must always be upper case, but the argument may be lower case. This commit safely converts the argument to upper case to easily match enum constants, in fact, this was already a feature that accidentally got removed with the addition of the asGeneric methods.