Replacer overhaul
Added Component support, updated the whole documentation of the class and updated the code itself to match new classes such as MCCollections. Also, numeric support will soon be removed from the Replacer class to be added as a pattern, which makes more sense. Right now, it still works when replacing on Strings, not on Components!