Remove some parametrization that wasn't needed
MCCommand, MCPlayer and MCConsole no longer use parametrized types. This is because if you are working on a platfrom, you already know the types you have to work with, for example, SpigotCommand obviously uses spigot player and console types, and if you are working platformless, then you still have a problem as you would still need to parametrize let's say MCCommandSender. if you work platformless, you will end up always using <?, ?>, this just removes the need of doing that...